Running benchmarks
I have added a new `harness` module to the GraceLibrary, which can be used to run Grace's benchmarks. A benchmark is simply a module that implements any number of methods ehose signatures begin with "benchmark".

To support the new benchmarking capibility, I've added a Python file that can be used to graph benchmarking results and also a new Bash script to automatic the run-render pipeline. Finally, I've updated the README to describe how to run and render benchmarks.